API: filter song list by multiple artists and albums
