+++ api/routes/filter-list.js
@@ -1,35 +1,12 @@
 const config = require('../../common/config');
 
-function getSortString(item) {
-    return item
-        .replace(/^the\s+/i, '')
-        .toLowerCase();
-}
-
-function sortCaseInsensitiveIgnorePrefix(prev, next) {
-    const prevSortString = getSortString(prev);
-    const nextSortString = getSortString(next);
-
-    if (prevSortString < nextSortString) {
-        return -1;
-    }
-
-    if (prevSortString > nextSortString) {
-        return 1;
-    }
-
-    return 0;
-}
+const { getInfoFilterQuery, sortCaseInsensitiveIgnorePrefix } = require('../helpers');
 
 function routeFilterList(key, subKey = null) {
     return async (req, res) => {
         let query = {};
         if (subKey && req.params[subKey]) {
-            const filter = req.params[subKey]
-                .split(',')
-                .map(item => ({ [`info.${subKey}`]: decodeURIComponent(item) }));
-
-            query = { $or: filter };
+            query = getInfoFilterQuery(req.params[subKey], subKey);
         }
 
         try {
